+++ src/ForbiddenStateError.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Bakame\Http\StructuredFields;
+
+use UnexpectedValueException;
+
+final class ForbiddenStateError extends UnexpectedValueException implements StructuredFieldError
+{
+}