+++ src/SerializationError.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Bakame\Http\StructuredFields;
-
-use UnexpectedValueException;
-
-final class SerializationError extends UnexpectedValueException implements StructuredFieldError
-{
-}