+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 - `OrderedList::from` named constructor which accepts a variadic list of members items
 - `Token::fromString` named constructor which accepts `string` and `Stringable` object
 - `Parameter::values` returns an array of all the values contained inside the `Parameters` instance
+- **[BC Break]** `ForbiddenStateError` to replace `SerializationError` 
 - **[BC Break]** `InnerList::fromList` to replace `InnerList::fromMembers`
 - **[BC Break]** `OrderedList::fromList` to replace `OrderedList::fromMembers`
 - **[BC Break]** `Parameter::value` to replace `InnerList::parameter` and `Item::parameter`
@@ -25,6 +26,7 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 - **[BC Break]** `OrderedList::__construct` is made private use `OrderedList::from` instead
 - **[BC Break]** `InnerList::__construct` is made private use `InnerList::fromList` instead
 - **[BC Break]** `Token::__construct` is made private use `Token::fromString` instead
+- **[BC Break]** `Parameter::get`, `Parameter::value`, `Parameter::pair` will throw `ForbiddenStateError` if the BareItem is in invalid state.
 
 ### Deprecated
 
@@ -42,6 +44,7 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 - **[BC Break]** `InnerList::parameters()` replaced by `InnerList::parameters` public readonly property
 - **[BC Break]** `InnerList::merge()` use `InnerList::push()` or `InnerList::unshift()` instead
 - **[BC Break]** `OrderedList::merge()` use `OrderedList::push()` or `OrderedList::unshift()` instead
+- **[BC Break]** `SerializationError` use `ForbiddenStateError` instead
 
 ## [0.1.0] - 2022-03-18